src/api_modelo_bento/requirements.txt
a2wsgi==1.10.10
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
astunparse==1.6.3
attrs==25.3.0
bentoml==1.4.24
cattrs==23.1.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-option-group==0.5.7
cloudpickle==3.1.1
colorama==0.4.6
duckdb==1.4.0
fastapi==0.116.2
flatbuffers==25.9.23
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
kantoku==0.18.3
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.6.4
namex==0.1.0
numpy==1.26.4
nvidia-ml-py==13.580.82
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-aiohttp-client==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
pip-requirements-parser==32.0.1
prometheus_client==0.22.1
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
schema==0.7.7
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.17.1
tensorflow-intel==2.17.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomli_w==1.2.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.17
uvicorn==0.35.0
watchfiles==1.1.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
xgboost==3.0.5
yarl==1.20.1
zipp==3.23.0
